community_contributions/morteza_joodaki/final_project/core/retriever.py
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity

class Retriever:
    def __init__(self, documents):
        self.texts = [doc["description"] for doc in documents]
        self.docs = documents
        self.vectorizer = TfidfVectorizer()
        self.matrix = self.vectorizer.fit_transform(self.texts)
    
    def search(self, query, top_k=5):
        q_vec = self.vectorizer.transform([query])
        scores = cosine_similarity(q_vec, self.matrix)[0]
        ranked = scores.argsort()[::-1][:top_k]
        return [self.docs[i] for i in ranked]