community_contributions/morteza_joodaki/final_project/requirements.txt
streamlit
python-dotenv
requests
numpy
scikit-learn
pandas